+++ packages/core/mashroom/src/config/mashroom_default_config.ts
@@ -18,7 +18,7 @@ const config: MashroomServerConfig = {
     }],
     ignorePlugins: [],
     indexPage: '/',
-    devModePreferredBuildTool: 'nx',
+    devModeDisableNxSupport: false,
     devModeNpmExecutionTimeoutSec: undefined,
 };