+++ packages/core/mashroom/type-definitions/api.ts
@@ -343,7 +343,7 @@ export type MashroomServerConfig = {
     readonly pluginPackageFolders: Array<PluginPackageFolder>;
     readonly ignorePlugins: Array<string>;
     readonly indexPage: string;
-    readonly devModePreferredBuildTool: 'npm' | 'nx' | null | undefined;
+    readonly devModeDisableNxSupport: boolean | null | undefined;
     readonly devModeNpmExecutionTimeoutSec: number | null | undefined;
 
     [key: string]: any;